// apps/web/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  ENM: 'ENM',
  CRF: 'CRF',
  DCS: 'DCS',
  MDS: 'MDS',
  ADJ: 'ADJ'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TypeApprenant = {
  ADJ: 'ADJ',
  CONCOURS_PRO: 'CONCOURS_PRO'
} as const

export type TypeApprenant = (typeof TypeApprenant)[keyof typeof TypeApprenant]


export const FonctionStage = {
  PARQUET: 'PARQUET',
  INSTRUCTION: 'INSTRUCTION',
  JE: 'JE',
  JAF: 'JAF',
  JAP: 'JAP',
  PENAL: 'PENAL',
  CIVIL: 'CIVIL',
  JCP: 'JCP'
} as const

export type FonctionStage = (typeof FonctionStage)[keyof typeof FonctionStage]


export const StatutEvaluation = {
  ATTENDUE: 'ATTENDUE',
  ENVOYEE: 'ENVOYEE',
  EN_COURS: 'EN_COURS',
  SOUMISE: 'SOUMISE',
  VALIDEE: 'VALIDEE',
  EN_RETARD: 'EN_RETARD'
} as const

export type StatutEvaluation = (typeof StatutEvaluation)[keyof typeof StatutEvaluation]


export const StatutStage = {
  PLANIFIE: 'PLANIFIE',
  EN_COURS: 'EN_COURS',
  TERMINE: 'TERMINE',
  CLOTURE: 'CLOTURE'
} as const

export type StatutStage = (typeof StatutStage)[keyof typeof StatutStage]